Agrosis_API-dev/requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
filelock==3.15.4
inflection==0.5.1
packaging==24.1
pillow==11.0.0
platformdirs==4.2.2
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
virtualenv==20.26.3
